Fix circular dependency by moving stuff to a new file
