+++ editor/src/components/inspector/common/control-status.ts
@@ -1,6 +1,5 @@
 import deepEqual from 'fast-deep-equal'
 //TODO: pass in colorTheme to utility functions to get rid of colorTheme here:
-import { colorTheme, UtopiaTheme } from '../../../uuiui/styles/theme'
 import type { JSXAttributes, StyleAttributeMetadata } from '../../../core/shared/element-template'
 import {
   modifiableAttributeIsAttributeNotFound,
@@ -9,10 +8,8 @@ import {
 } from '../../../core/shared/element-template'
 import type { GetModifiableAttributeResult } from '../../../core/shared/jsx-attributes'
 import { getModifiableJSXAttributeAtPath } from '../../../core/shared/jsx-attributes'
-import { isLeft, isRight, Either, right, defaultEither } from '../../../core/shared/either'
-import Utils from '../../../utils/utils'
+import { isLeft, isRight, right, defaultEither } from '../../../core/shared/either'
 import type { ParsedCSSProperties, ParsedPropertiesKeys } from './css-utils'
-import { CSSNumber } from './css-utils'
 import type { MultiselectAtProps, MultiselectAtStringProps } from './property-path-hooks'
 import * as PP from '../../../core/shared/property-path'
 import type { IcnColor } from '../../../uuiui'
@@ -64,30 +61,6 @@ export type ControlStatus =
   | 'multiselect-disabled' // at least one element in the multi-selection is 'disabled'
   | 'multiselect-trivial-default' // this single-selected element's property is detected from measurement and we think its not important to show it to the user
 
-const AllControlStatuses: Array<ControlStatus> = [
-  'off',
-  'simple',
-  'simple-unknown-css',
-  'unset',
-  'disabled',
-  'unoverwritable',
-  'controlled',
-  'detected-fromcss',
-  'detected',
-  'trivial-default',
-  'overridden',
-  'multiselect-identical-simple',
-  'multiselect-simple-unknown-css',
-  'multiselect-identical-unset',
-  'multiselect-mixed-simple-or-unset',
-  'multiselect-controlled',
-  'multiselect-detected-fromcss',
-  'multiselect-detected',
-  'multiselect-unoverwritable',
-  'multiselect-disabled',
-  'multiselect-trivial-default',
-]
-
 export function isControlledStatus(controlStatus: ControlStatus): boolean {
   switch (controlStatus) {
     case 'controlled':
@@ -117,156 +90,6 @@ export function isControlledStatus(controlStatus: ControlStatus): boolean {
       const _exhaustiveCheck: never = controlStatus
       throw new Error(`Unhandled status ${JSON.stringify(controlStatus)}`)
   }
-}
-
-const controlStylesByStatus: { [key: string]: ControlStyles } = Utils.mapArrayToDictionary(
-  AllControlStatuses,
-  (status: ControlStatus) => status,
-  (status: ControlStatus): ControlStyles => {
-    let fontStyle = 'normal'
-    let fontWeight = 400
-    let mainColor: string = colorTheme.fg1.value
-    let secondaryColor: string = colorTheme.fg7.value
-
-    let borderColor: string = 'transparent'
-    // text inputs
-    let backgroundColor: string = colorTheme.bg2.value
-    let focusedBackgroundColor: string = colorTheme.bg0.value
-    let segmentSelectorColor: string = colorTheme.bg3.value
-    let trackColor = colorTheme.fg7.value
-    let railColor = colorTheme.bg3.value
-    let strokePrimaryColor = colorTheme.fg5.value
-    let strokeSecondaryColor = colorTheme.fg7.value
-    let strokeTertiaryColor = colorTheme.fg9.value
-    let set = true
-    let interactive = true
-    let mixed = false
-    let unknown = false
-    let showContent = true
-    let unsettable = true
-    let iconColor: IcnColor = 'main'
-
-    switch (status) {
-      case 'simple':
-      case 'multiselect-identical-simple':
-        break
-      case 'detected':
-      case 'multiselect-detected':
-      case 'unset':
-      case 'multiselect-identical-unset':
-        set = false
-        unsettable = false
-        // this makes it interactive if the component provides these values
-        mainColor = 'var(--control-styles-interactive-unset-main-color)'
-        secondaryColor = 'var(--control-styles-interactive-unset-secondary-color)'
-        trackColor = 'var(--control-styles-interactive-unset-track-color)'
-        railColor = 'var(--control-styles-interactive-unset-rail-color)'
-        break
-      case 'controlled':
-      case 'multiselect-controlled':
-      case 'multiselect-mixed-simple-or-unset':
-        interactive = true
-        mainColor = colorTheme.dynamicBlue.value
-        secondaryColor = colorTheme.dynamicBlue.value
-        trackColor = colorTheme.dynamicBlue.value
-        strokePrimaryColor = colorTheme.dynamicBlue.value
-        showContent = true
-        break
-      case 'trivial-default':
-      case 'multiselect-trivial-default':
-        interactive = true
-        showContent = false
-        mainColor = colorTheme.fg7.value
-        secondaryColor = colorTheme.fg7.value
-        trackColor = colorTheme.bg5.value
-        railColor = colorTheme.bg3.value
-        break
-      case 'detected-fromcss':
-      case 'multiselect-detected-fromcss':
-        mainColor = colorTheme.css.value
-        secondaryColor = colorTheme.css.value
-        trackColor = colorTheme.css.value
-        break
-      case 'simple-unknown-css':
-      case 'multiselect-simple-unknown-css':
-        unknown = true
-        interactive = true
-        mainColor = 'orange'
-        secondaryColor = 'orange'
-        trackColor = 'orange'
-        backgroundColor = 'transparent'
-        break
-      case 'unoverwritable':
-      case 'multiselect-unoverwritable':
-        interactive = false
-        mainColor = colorTheme.fg8.value
-        secondaryColor = colorTheme.fg8.value
-        trackColor = colorTheme.fg8.value
-        railColor = colorTheme.fg8.value
-        unsettable = false
-        break
-
-      case 'off':
-        set = false
-        interactive = false
-        mainColor = colorTheme.fg7.value
-        secondaryColor = colorTheme.fg7.value
-        trackColor = colorTheme.fg7.value
-        showContent = false
-        unsettable = false
-        break
-      case 'disabled':
-      case 'multiselect-disabled':
-        interactive = false
-        mainColor = colorTheme.fg7.value
-        secondaryColor = colorTheme.fg7.value
-        backgroundColor = 'transparent'
-        segmentSelectorColor = 'transparent'
-        trackColor = 'transparent'
-        showContent = true
-        unsettable = false
-        break
-      case 'overridden':
-        interactive = true
-        mainColor = colorTheme.brandNeonPink.value
-        borderColor = colorTheme.brandNeonPink.value
-        secondaryColor = colorTheme.primary.value
-        trackColor = colorTheme.primary.value
-        strokePrimaryColor = colorTheme.primary.value
-        showContent = true
-        iconColor = 'overridden'
-        break
-      default:
-        break
-    }
-
-    return {
-      fontStyle,
-      fontWeight,
-      mainColor,
-      secondaryColor,
-      borderColor,
-      backgroundColor,
-      focusedBackgroundColor,
-      segmentSelectorColor,
-      strokePrimaryColor,
-      strokeSecondaryColor,
-      strokeTertiaryColor,
-      trackColor,
-      railColor,
-      set,
-      interactive,
-      mixed,
-      showContent,
-      unknown,
-      unsettable,
-      iconColor,
-    }
-  },
-)
-
-export function getControlStyles(controlStatus: ControlStatus): ControlStyles {
-  return controlStylesByStatus[controlStatus]
 }
 
 interface SinglePropertyStatus {
