+++ editor/src/components/inspector/common/longhand-shorthand-hooks.ts
@@ -14,7 +14,8 @@ import {
 import { useDispatch } from '../../editor/store/dispatch-context'
 import { useEditorState } from '../../editor/store/store-hook'
 import type { PropertyStatus } from './control-status'
-import { getControlStatusFromPropertyStatus, getControlStyles } from './control-status'
+import { getControlStatusFromPropertyStatus } from './control-status'
+import { getControlStyles } from './control-styles'
 import type { ParsedProperties, ParsedPropertiesKeys } from './css-utils'
 import { printCSSValue } from './css-utils'
 import type { ReadonlyRef } from './inspector-utils'
