+++ editor/src/components/inspector/common/property-controls-hooks.ts
@@ -43,8 +43,8 @@ import {
   calculatePropertyStatusForSelection,
   ControlStatus,
   getControlStatusFromPropertyStatus,
-  getControlStyles,
 } from './control-status'
+import { getControlStyles } from './control-styles'
 import {
   keepDeepReferenceEqualityIfPossible,
   useKeepReferenceEqualityIfPossible,
