+++ editor/src/components/inspector/controls/select-control.tsx
@@ -6,7 +6,8 @@ import CreatableSelect from 'react-select/creatable'
 import type { IndicatorProps } from 'react-select/src/components/indicators'
 import Utils from '../../../utils/utils'
 import type { DEPRECATEDControlProps, DEPRECATEDGenericControlOptions } from './control'
-import { getControlStyles, ControlStatus, isControlledStatus } from '../common/control-status'
+import { ControlStatus, isControlledStatus } from '../common/control-status'
+import { getControlStyles } from '../common/control-styles'
 import type { ValueType } from 'react-select/src/types'
 import { PortalTargetID } from '../../../core/shared/utils'
 import type { IcnProps } from '../../../uuiui'
