+++ editor/src/components/inspector/controls/url-background-layer-metadata-controls.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import * as PP from '../../../core/shared/property-path'
 import type { CSSURLFunctionBackgroundLayer, CSSBackgroundLayers } from '../common/css-utils'
 import { isCSSImageURLBackgroundLayer } from '../common/css-utils'
-import { getControlStyles } from '../common/control-status'
+import { getControlStyles } from '../common/control-styles'
 import { UIGridRow } from '../widgets/ui-grid-row'
 import { PropertyLabel } from '../widgets/property-label'
 import type { BackgroundLayerControlsProps } from '../sections/style-section/background-subsection/background-picker'
