+++ editor/src/components/inspector/sections/image-section/image-density-control.tsx
@@ -3,7 +3,7 @@ import type { ElementPath } from '../../../../core/shared/project-file-types'
 import type { EditorAction, EditorDispatch } from '../../../editor/action-types'
 import { showToast, updateFrameDimensions } from '../../../editor/actions/action-creators'
 import { OptionChainControl } from '../../controls/option-chain-control'
-import { getControlStyles } from '../../common/control-status'
+import { getControlStyles } from '../../common/control-styles'
 import { notice } from '../../../common/notice'
 import type { Size } from '../../../../core/shared/math-utils'
 import { size } from '../../../../core/shared/math-utils'
