+++ editor/src/components/inspector/sections/layout-section/layout-system-subsection/layout-system-controls.tsx
@@ -29,8 +29,8 @@ import {
   ControlStatus,
   ControlStyles,
   getControlStatusFromPropertyStatus,
-  getControlStyles,
 } from '../../../common/control-status'
+import { getControlStyles } from '../../../common/control-styles'
 import type { CanvasControlWithProps } from '../../../common/inspector-atoms'
 import { useControlModeWithCycle } from '../../../common/inspector-utils'
 import { useInspectorInfoLonghandShorthand } from '../../../common/longhand-shorthand-hooks'
