+++ editor/src/components/inspector/sections/style-section/text-subsection/text-auto-sizing-control.tsx
@@ -6,7 +6,7 @@ import { applyCommandsAction } from '../../../../editor/actions/action-creators'
 import { useDispatch } from '../../../../editor/store/dispatch-context'
 import { Substores, useEditorState, useRefEditorState } from '../../../../editor/store/store-hook'
 import type { ControlStatus } from '../../../common/control-status'
-import { getControlStyles } from '../../../common/control-status'
+import { getControlStyles } from '../../../common/control-styles'
 import { cssNumber } from '../../../common/css-utils'
 import type { OptionChainOption } from '../../../controls/option-chain-control'
 import { OptionChainControl } from '../../../controls/option-chain-control'
