+++ editor/src/components/inspector/sections/style-section/text-subsection/text-subsection.tsx
@@ -43,7 +43,7 @@ import { PropertyRow } from '../../../widgets/property-row'
 import { FontFamilySelect } from './font-family-select'
 import { FontVariantSelect } from './font-variant-select'
 import { FlexRow } from 'utopia-api'
-import { getControlStyles } from '../../../common/control-status'
+import { getControlStyles } from '../../../common/control-styles'
 import { Utils } from '../../../../../uuiui-deps'
 import { TextRelatedProperties } from '../../../../../core/properties/css-properties'
 import { useContextSelector } from 'use-context-selector'
