+++ editor/src/components/navigator/dependency-list-input-field.tsx
@@ -7,7 +7,7 @@ import Utils from '../../utils/utils'
 import type { SelectOption } from '../inspector/controls/select-control'
 import { SelectControl } from '../inspector/controls/select-control'
 import { top1000NPMPackagesOptions } from './top1000NPMPackagesOptions'
-import { getControlStyles } from '../inspector/common/control-status'
+import { getControlStyles } from '../inspector/common/control-styles'
 import { Tooltip, StringInput, OnClickOutsideHOC, UtopiaTheme, useColorTheme } from '../../uuiui'
 
 interface DependencySearchSelectProps {
