+++ editor/src/uuiui/inputs/string-input.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled'
 import composeRefs from '@seznam/compose-react-refs'
 import React from 'react'
 import type { ControlStatus, ControlStyles } from '../../components/inspector/common/control-status'
-import { getControlStyles } from '../../components/inspector/common/control-status'
+import { getControlStyles } from '../../components/inspector/common/control-styles'
 import { preventDefault, stopPropagation } from '../../components/inspector/common/inspector-utils'
 import { useColorTheme, UtopiaTheme } from '../styles/theme'
 import { InspectorInput, InspectorInputEmotionStyle } from './base-input'
